requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.3
mysql-connector-python==8.2.0
scikit-learn==1.3.2
statsmodels==0.14.0
python-dateutil==2.8.2
pydantic==2.5.0
